refactor: extract Zoom manifest to separate file, modernize types
Maintainability and code quality improvements:

- Extract Manifest() from zoom.go (477 lines) into manifest.go (295 lines),
  following the pattern established by the Google connector
- zoom.go is now 189 lines — focused on connector struct, HTTP client,
  and error handling
- Replace map[string]interface{} with map[string]any in create/update
  meeting params and request types (modern Go style)
- Verified all 51 tests use t.Parallel() — no race conditions
- Each test creates its own httptest.NewServer and connector — no shared
  mutable state
